# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Website Builder Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List all testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"

    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create testimonial (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimonialBody"
      responses:
        "201":
          description: Testimonial created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update testimonial (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestimonialBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"

    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete testimonial (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        username:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - fullName
        - phone
        - email
        - username
        - role
        - createdAt
...
    Testimonial:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        comment:
          type: string
        rating:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clientName
        - comment
        - rating
        - isActive
        - createdAt
    CreateTestimonialBody:
      type: object
      properties:
        clientName:
          type: string
        comment:
          type: string
        rating:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - clientName
        - comment
        - rating
    UpdateTestimonialBody:
      type: object
      properties:
        clientName:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
...