# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Website Builder Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  clientName: string;
  comment: string;
  rating: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}